pyproject.toml
[project]
name = "d5reports"
version = "0.1.0"
description = "Internal reporting system for Division5"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "argon2-cffi>=23.1.0",
    "loguru>=0.7.2",
    "pymysql>=1.1.1",
    "python-ulid>=3.0.0",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.7.2",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 79
indent-width = 4

target-version = "py312"

[tool.ruff.lint.isort]
force-sort-within-sections = true
combine-as-imports = true
length-sort = true


[tool.coverage.run]
source = ["./tests"]
omit = ['*/__init__.py', '*/tests/*', '*/migrations/*', '*/alembic/*', '*/enums/*']

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
